Replace Dell string types with DisplayString
Dell uses a number of improperly specified strings syntaxes in the
IDRAC-MIB-SMIv2. Values using these syntaxses are string values, but the syntax
is declared as an octet string, so these values will get displayed
as hex strings. See `systemBIOSManufacturerName` in the following output:

    systemBIOSManufacturerName{instance="10.2.11.112", job="idrac_snmp",
    systemBIOSIndex="1", systemBIOSManufacturerName="0x44656C6C20496E632E",
    systemBIOSchassisIndex="1"}

This commit adds a sed script to build-mibs-db.sh that replaces these types
with a `DisplayString` variant, allowing them to display correctly in
prometheus metrics.